Add ecs_e7 lunch target
The ECS Cruise E7 board is a near-clone of the MRD7, and so is the
configuration.  This is just a copy with the product strings changed;
the only significant change is that the camera modules loaded in
init.atomisp.sh are different.

Other changes will come later in future commits.

Change-Id: I9f521370be1679f78c4ef925cdca4e093b563cb2
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>