[WORKAROUND] minnowboard-max: Leave display power well ON after HDMI cable is plugged out
Apparently, there is a known issue on ValleyView, regarding the hotplug
detection after the display outputs are disabled. If the display power
well is disabled, the IRQ steering logic inside the chip is also
disabled. That means that no interrupts will be generated when the cable
is plugged back in, hence the display is useless until the next boot.

This commit will force the display power well to be always on. This is
temporary, until a better SW fix is found.

Change-Id: I8312fb0d64bab607c66a2d69036afac1b2e45693
Signed-off-by: Laurentiu Palcu <laurentiu.palcu@intel.com>
Reviewed-on: https://android.intel.com/285888
Reviewed-by: Dolca, Robert <robert.dolca@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Pircalabu, Petre Ovidiu <petre.ovidiu.pircalabu@intel.com>